fixed map values again

+++ src/main/java/frc/robot/RobotMap.java
@@ -7,7 +7,7 @@ public class RobotMap {
      * All constants should be ordered by subsystem/use
      * The units of each measurement should be specified in a comment
      */
-    public static final double HATCH_PIVOT_MOTOR = 1; // Port number (dummy value)
+    public static final int HATCH_PIVOT_MOTOR = 1; // Port number (dummy value)
     public static final double HATCH_P_VALUE = 1;
     public static final double HATCH_I_VALUE = 1;
     public static final double HATCH_D_VALUE = 1;
